Issue #154 refactor: model 패키지 리팩토링
