+++ app/src/main/java/com/lateinit/rightweight/data/database/AppPreferencesDataStore.kt
@@ -7,7 +7,7 @@ import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import androidx.datastore.preferences.preferencesDataStore
 import com.google.gson.Gson
-import com.lateinit.rightweight.data.model.User
+import com.lateinit.rightweight.data.model.local.User
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
