+++ app/src/main/java/com/lateinit/rightweight/data/datasource/local/impl/UserLocalDataSourceImpl.kt
@@ -5,7 +5,7 @@ import com.lateinit.rightweight.data.database.dao.UserDao
 import com.lateinit.rightweight.data.database.intermediate.HistoryWithHistoryExercises
 import com.lateinit.rightweight.data.database.intermediate.RoutineWithDays
 import com.lateinit.rightweight.data.datasource.local.UserLocalDataSource
-import com.lateinit.rightweight.data.model.User
+import com.lateinit.rightweight.data.model.local.User
 import kotlinx.coroutines.flow.Flow
 import java.time.LocalDate
 import javax.inject.Inject
