+++ app/src/main/java/com/lateinit/rightweight/data/datasource/remote/impl/LoginDataSourceImpl.kt
@@ -3,8 +3,8 @@ package com.lateinit.rightweight.data.datasource.remote.impl
 import com.lateinit.rightweight.data.AuthApiService
 import com.lateinit.rightweight.data.LoginResponse
 import com.lateinit.rightweight.data.datasource.remote.LoginDataSource
-import com.lateinit.rightweight.data.model.LoginRequestBody
-import com.lateinit.rightweight.data.model.PostBody
+import com.lateinit.rightweight.data.model.remote.LoginRequestBody
+import com.lateinit.rightweight.data.model.remote.PostBody
 import javax.inject.Inject
 
 class LoginDataSourceImpl @Inject constructor(
