+++ app/src/main/java/com/lateinit/rightweight/data/datasource/remote/impl/UserRemoteDataSourceImpl.kt
@@ -2,18 +2,18 @@ package com.lateinit.rightweight.data.datasource.remote.impl
 
 import com.lateinit.rightweight.data.UserApiService
 import com.lateinit.rightweight.data.datasource.remote.UserRemoteDataSource
-import com.lateinit.rightweight.data.model.Direction
-import com.lateinit.rightweight.data.model.DocumentResponse
-import com.lateinit.rightweight.data.model.FiledReferenceData
-import com.lateinit.rightweight.data.model.FilterData
-import com.lateinit.rightweight.data.model.FilterOperator
-import com.lateinit.rightweight.data.model.FromData
-import com.lateinit.rightweight.data.model.OrderByData
-import com.lateinit.rightweight.data.model.RunQueryBody
-import com.lateinit.rightweight.data.model.StructuredQueryData
-import com.lateinit.rightweight.data.model.WhereData
-import com.lateinit.rightweight.data.model.WriteModelData
-import com.lateinit.rightweight.data.model.WriteRequestBody
+import com.lateinit.rightweight.data.model.remote.Direction
+import com.lateinit.rightweight.data.model.remote.DocumentResponse
+import com.lateinit.rightweight.data.model.remote.FiledReferenceData
+import com.lateinit.rightweight.data.model.remote.FilterData
+import com.lateinit.rightweight.data.model.remote.FilterOperator
+import com.lateinit.rightweight.data.model.remote.FromData
+import com.lateinit.rightweight.data.model.remote.OrderByData
+import com.lateinit.rightweight.data.model.remote.RunQueryBody
+import com.lateinit.rightweight.data.model.remote.StructuredQueryData
+import com.lateinit.rightweight.data.model.remote.WhereData
+import com.lateinit.rightweight.data.model.remote.WriteModelData
+import com.lateinit.rightweight.data.model.remote.WriteRequestBody
 import com.lateinit.rightweight.data.remote.model.HistoryField
 import com.lateinit.rightweight.data.remote.model.RootField
 import com.lateinit.rightweight.data.remote.model.RoutineField
