+++ app/src/main/java/com/lateinit/rightweight/data/model/remote/DataValue.kt
@@ -5,26 +5,21 @@ import com.google.gson.annotations.SerializedName
 sealed class ValueData
 
 data class StringValue(
-    @field:SerializedName("stringValue")
+    @SerializedName("stringValue")
     val value: String? = null
 ): ValueData()
 
-data class BooleanValue(
-    @field:SerializedName("booleanValue")
-    val value: Boolean? = null
-): ValueData()
-
 data class IntValue(
-    @field:SerializedName("integerValue")
+    @SerializedName("integerValue")
     val value: String? = null
 ): ValueData()
 
 data class TimeStampValue(
-    @field:SerializedName("timestampValue")
+    @SerializedName("timestampValue")
     val value: String? = null
 ): ValueData()
 
 data class MapValue(
-    @field:SerializedName("mapValue")
+    @SerializedName("mapValue")
     val value: MapValueRootField? = null
 ): ValueData()
