+++ app/src/main/java/com/lateinit/rightweight/data/model/remote/Documents.kt
@@ -1,8 +1,7 @@
-package com.lateinit.rightweight.data.model
+package com.lateinit.rightweight.data.model.remote
 
 data class DocumentResponse<T>(val document: DetailResponse<T>?)
 
 data class DocumentsResponse<T>(val documents: List<DetailResponse<T>>?)
 
 data class DetailResponse<T> (val name: String, val fields: T)
-
