+++ app/src/main/java/com/lateinit/rightweight/data/model/remote/RunQueryModel.kt
@@ -1,4 +1,4 @@
-package com.lateinit.rightweight.data.model
+package com.lateinit.rightweight.data.model.remote
 
 import com.lateinit.rightweight.data.remote.model.ValueData
