--- app/src/main/java/com/lateinit/rightweight/data/remote/model/sharedRoutine/SharedRoutineResponse.kt
@@ -1,8 +1,0 @@
-package com.lateinit.rightweight.data.remote.model.sharedRoutine
-
-import com.google.gson.annotations.SerializedName
-
-data class SharedRoutineResponse(
-    @field:SerializedName("documents")
-    val documents: List<SharedRoutineItem?>? = null
-)
