+++ app/src/main/java/com/lateinit/rightweight/data/repository/SharedRoutineRepository.kt
@@ -3,8 +3,8 @@ package com.lateinit.rightweight.data.repository
 import androidx.paging.PagingData
 import com.lateinit.rightweight.data.database.entity.SharedRoutine
 import com.lateinit.rightweight.data.database.intermediate.SharedRoutineWithDays
-import com.lateinit.rightweight.data.model.SharedRoutineSortType
-import com.lateinit.rightweight.data.model.WriteModelData
+import com.lateinit.rightweight.data.model.remote.SharedRoutineSortType
+import com.lateinit.rightweight.data.model.remote.WriteModelData
 import kotlinx.coroutines.flow.Flow
 
 interface SharedRoutineRepository {
