+++ app/src/main/java/com/lateinit/rightweight/ui/routine/detail/RoutineDetailViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lateinit.rightweight.data.model.UpdateData
-import com.lateinit.rightweight.data.model.WriteModelData
+import com.lateinit.rightweight.data.model.remote.UpdateData
+import com.lateinit.rightweight.data.model.remote.WriteModelData
 import com.lateinit.rightweight.data.repository.RoutineRepository
 import com.lateinit.rightweight.data.repository.SharedRoutineRepository
 import com.lateinit.rightweight.data.repository.UserRepository
